# swagger.yaml
...
openapi: '3.0.3'
info:
  title: "Groupomania API"
  description: "List of all routes availables for the API"
  version: "1.0"
produces: ["application/json"]
servers:
  - url: "http://localhost:3000/api"
    description: "Local server test API"
...
paths:
...
  /post:
    post:
      summary: "Create new post"
      description: "Create new post in community"
      tags: ["Post"]
      security:
        - bearerAuth: []
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: "string"
                    example: "Mon super post"
                  content:
                    type: "string"
                    example: "Contenu de mon super post"
                  communityId:
                    type: "integer"
                    example: 1
                  shareFromPostId:
                    type: integer
                    example: 1
                    description: If is a shared post, specify the id of the parent post
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/ApiSuccessResponse'
              example:
                code: 200
                data: {}
                success: true
        401:
          description: "Not logged-in"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 401
                errorMessage: "Vous n'êtes pas connecté. Merci de vous authentifier."
                error: {}
                data: {}
                success: false
        500:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 500
                errorMessage: "Internal server error"
                error: {}
                data: {}
                success: false
    get:
      summary: "Read user feed"
      description: "Read all post from followed community, or other random suggestion."
      tags: ["Post"]
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
          description: The ID of the user to get posts from
        - in: query
          name: limit
          schema:
            type: integer
          description: The max number of posts to return (max = 100 per request)
          example: 20
        - in: query
          name: minPostId
          schema:
            type: integer
          description: Get only results with more than the specified id
        - in: query
          name: maxPostId
          schema:
            type: integer
          description: Get only results with less than the specified id
        - in: query
          name: favorite
          schema:
            type: boolean
          description: Get only results from current loggedin user favorites

      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/ApiSuccessResponse'
              example:
                code: 200
                data:
                  posts:
                    -
                      id: 3
                      title: "Ma belle communauté"
                      content: "Voici donc mon histoire, qui raconte un super contenu qui n'en finira jamais... Bonne nuit!"
                      likes: 2
                      comments: 0
                      createdAt: "2021-10-21T16:27:07.000Z"
                      updatedAt: "2021-10-25T08:53:10.000Z"
                      CommunityId: 18
                      ShareFromPostId: null
                      UserId: 1
                      PostFiles: []
                      PostComments: []
                      PostLikes: []
        500:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 500
                errorMessage: "Internal server error"
                error: {}
                data: {}
                success: false
  /post/{id}:
    get:
      summary: "Read a post"
      description: "Read a specific post"
      tags: ["Post"]
      parameters:
        - name: id
          description: The post ID to read
          in: path
          required: true
          type: integer
          example: 1
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/ApiSuccessResponse'
              example:
                code: 200
                data:
                  post:
                    id: 3
                    title: "Ma belle communauté"
                    content: "Voici donc mon histoire, qui raconte un super contenu qui n'en finira jamais... Bonne nuit!"
                    likes: 2
                    comments: 0
                    createdAt: "2021-10-21T16:27:07.000Z"
                    updatedAt: "2021-10-25T08:53:10.000Z"
                    CommunityId: 18
                    ShareFromPostId: null
                    UserId: 1
                    PostFiles: []
                    PostComments:
                      - 
                        id: 2
                        content:  "Ceci est vraiment un SUPERBE poste de FIFOU!"
                        like: 1
                        createdAt: "2021-10-21T18:24:30.000Z"
                        updatedAt: "2021-10-25T09:09:18.000Z"
                        PostId: 3
                        UserId: 2
                        PostCommentId: null
                    PostLikes:
                      - 
                        id: 2
                        createdAt: "2021-10-21T17:32:44.000Z"
                        updatedAt: "2021-10-21T17:32:44.000Z"
                        PostId: 3
                        UserId: 2
                success: true
        500:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 500
                errorMessage: "Internal server error"
                error: {}
                data: {}
                success: false
    put:
      summary: "Update a post"
      description: "Update a specific post in community"
      tags: ["Post"]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: "string"
                  description: "The post title"
                  example: "Mon super titre"
                content:
                  type: "string"
                  description: "The post content"
                  example: "Voici donc une super histoire qui n'en finira jamais... ! :)"
      parameters:
        - name: id
          description: The post ID to update
          in: path
          required: true
          type: integer
          example: 1
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/ApiSuccessResponse'
              example:
                code: 200
                data: {}
                success: true
        401:
          description: "Not logged-in"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 401
                errorMessage: "Vous n'êtes pas connecté. Merci de vous authentifier."
                error: {}
                data: {}
                success: false
        403:
          description: "Access denied"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 403
                errorMessage: "Vous n'avez pas la permission d'accéder à ce contenu."
                error: {}
                data: {}
                success: false
        500:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 500
                errorMessage: "Internal server error"
                error: {}
                data: {}
                success: false
    delete:
      summary: "Delete a post"
      description: "Delete a post from specific community"
      tags: ["Post"]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          description: The post ID to delete
          in: path
          required: true
          type: integer
          example: 1
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/ApiSuccessResponse'
              example:
                code: 200
                data: {}
                success: true
        401:
          description: "Not logged-in"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 401
                errorMessage: "Vous n'êtes pas connecté. Merci de vous authentifier."
                error: {}
                data: {}
                success: false
        403:
          description: "Access denied"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 403
                errorMessage: "Vous n'avez pas la permission d'accéder à ce contenu."
                error: {}
                data: {}
                success: false
        500:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 500
                errorMessage: "Internal server error"
                error: {}
                data: {}
                success: false
...
  /post/{id}/files:
    get:
      summary: "Read files"
      description: "Read all files from a specific post"
      tags: ["Post"]
      parameters:
        - name: id
          description: The post ID to read files
          in: path
          required: true
          type: integer
          example: 1
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/ApiSuccessResponse'
              example:
                code: 200
                data:
                  files: []
                success: true
        500:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 500
                errorMessage: "Internal server error"
                error: {}
                data: {}
                success: false
  /post/{id}/like:
    post:
      summary: "Like / Dislike a post"
      description: "Like / Dislike a specific post"
      tags: ["Post"]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          description: The post ID to like / dislike
          in: path
          required: true
          type: integer
          example: 1
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  like:
                    type: "integer"
                    format: "int32"
                    examples:
                      like:
                        value: 1
                      dislike:
                        value: -1
                      cancel:
                        value: 0
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/ApiSuccessResponse'
                  
        401:
          description: "Not logged-in"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 401
                errorMessage: "Vous n'êtes pas connecté. Merci de vous authentifier."
                error: {}
                data: {}
                success: false
        500:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: '#/components/ApiErrorResponse'
              example:
                code: 500
                errorMessage: "Internal server error"
                error: {}
                data: {}
                success: false
...
components:
...
  securitySchemes:
    bearerAuth:           # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
...